[fixes #44633429] Creating a mixin for 'transfer'
Implementing a common mixin for transfer related
action like TransferPlatesToPlates and
TransferTubesToTubes.
