+++ lib/lims-core/actions/transfer_tubes_to_tubes.rb
@@ -1,4 +1,5 @@
 require 'lims/core/actions/action'
+require 'lims/core/actions/transfer_action'
 
 require 'lims/core/laboratory/spin_column'
 require 'lims/core/laboratory/tube'
@@ -16,56 +17,14 @@ module Lims::Core
     # Type is the type of the aliquot.
     class TransferTubesToTubes
       include Action
-
-      attribute :transfers, Array, :required => true, :writer => :private
-
-      def _validate_parameters
-        raise ArgumentError, "The transfer array should not be null." unless transfers
-        raise ArgumentError, "You should give the fraction OR the amount of the transfer, not both." unless valid_amount_and_fraction
-      end
-
-      def valid_amount_and_fraction
-        valid = true
-        transfers.each do |transfer|
-          if (transfer["fraction"].nil? && transfer["amount"].nil?) || (transfer["fraction"] && transfer["amount"])
-            valid = false
-            break
-          end
-        end
-        valid
-      end
+      include TransferAction
 
       # transfer the given fraction of aliquot from tube-like asset(s)
       # to tube-like asset(s)
       def _call_in_session(session)
-        sources = []
-        targets = []
-        amounts = []
-        transfers.each do |transfer|
-          fraction = transfer["fraction"]
-          if fraction
-            amounts << transfer["source"].quantity * fraction
-          else
-            amounts << transfer["amount"]
-          end
-        end
 
-        transfers.zip(amounts) do |transfer, amount|
-          source = transfer["source"]
-          target = transfer["target"]
-          target << source.take_amount(amount)
+        _transfer(transfers, _amounts(transfers))
 
-          unless transfer["aliquot_type"].nil?
-            target.each do |aliquot|
-              aliquot.type = transfer["aliquot_type"]
-            end
-          end
-
-          sources << source
-          targets << target
-        end
-
-        { :sources => sources.uniq, :targets => targets.uniq}
       end
     end
   end
